Lecture: replace link to ipynb (#282)
In our challenges, resolving the images and attachments doesn't work properly. For this reason, we have to use instead of local links here the links in the GitHub repo. 

In order for the images to be resolved correctly, we need the "raw" URL here: The image should be embedded as such in the website - the "normal" URL however is the preview on GitHub. For other files, we should use the "normal" URL so that the preview is displayed.